{{cookiecutter.extension_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.extension_name}}"
version = "{{ cookiecutter.version }}"
description = "{{ cookiecutter.description }}"
authors = ["{{ cookiecutter.author }}"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
biscuit-editor = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
